[#18] refactor - 프로젝트 구조에 맞게 UserPage refactoring
