+++ client/src/components/index.ts
@@ -1,6 +1,8 @@
 export { default as ProductCard } from './small/ProductCard';
 export { default as SectionDivider } from './small/SectionDivider';
 export { default as SectionHeader } from './small/SectionHeader';
+export { default as UserHeader } from './small/UserHeader';
+export { default as UserFooter } from './small/UserFooter';
 
 export { default as CategoryIconGrid } from './medium/CategoryIconGrid';
 export { default as ProductCardGrid } from './medium/ProductCardGrid';
@@ -14,3 +16,5 @@ export { default as EatNowSection } from './large/EatNowSection';
 export { default as ForYouSection } from './large/ForYouSection';
 export { default as PageHeader } from './large/PageHeader';
 export { default as HotDealSection } from './large/BestSellerSection';
+export { default as JoinSection } from './large/JoinSection';
+export { default as LoginSection } from './large/LoginSection';
